requirements.txt
appdirs==1.4.3
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
entrypoints==0.3
flake8==3.7.8
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
mccabe==0.6.1
mecab-python3==0.996.2
numpy==1.17.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
requests==2.22.0
six==1.12.0
stanfordnlp==0.2.0
toml==0.10.0
torch==1.2.0
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
yapf==0.28.0
